tests/wpt/tests/jpegxl/modular-lossy.html

<!DOCTYPE html>
<title>JPEG XL: modular lossy</title>
<link rel="help" href="https://github.com/web-platform-tests/interop-jpegxl">
<meta name="assert" content="Decodes lossy JPEG XL fixture.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script>

promise_test(async () => {
  const img = await loadImage('resources/3x3_srgb_lossy.jxl');
  assert_equals(img.naturalWidth, 3);
  assert_equals(img.naturalHeight, 3);
}, 'Lossy JPEG XL fixture decodes.');

</script>
